🚚 Rename path to the LKPR sample app component
